--- web/src/beta/lib/unsafeBuiltinPlugins/index.ts
@@ -1,5 +1,0 @@
-import { loadPlugins } from "./load";
-
-export * from "./types";
-
-export default await loadPlugins();
