+++ web/src/services/config/unsafeBuiltinPlugin.ts
@@ -1,15 +1,36 @@
-import type { UnsafeBuiltinPlugin } from "../../beta/lib/unsafeBuiltinPlugins";
+import { FC } from "react";
 
-export type { UnsafeBuiltinPlugin } from "../../beta/lib/unsafeBuiltinPlugins";
+export type UnsafeBuiltinPlugin = {
+  id: string;
+  name: string;
+  widgets?: UnsafeBuiltinWidget[];
+  blocks?: UnsafeBuiltinBlock[];
+};
 
-export async function loadUnsafeBuiltinPlugins() {
-  try {
-    const unsafeBuiltinPlugins = (
-      await import(/* @vite-ignore */ "src/beta/lib/unsafeBuiltinPlugins")
-    ).default as UnsafeBuiltinPlugin[];
-    return unsafeBuiltinPlugins;
-  } catch (e) {
-    console.error("unsafe builtin plugin load failed", e);
-  }
-  return undefined;
-}
+type UnsafeBuiltinWidget = UnsafeBuiltinPluginExtension<"widget">;
+
+type UnsafeBuiltinBlock = UnsafeBuiltinPluginExtension<"block">;
+
+type UnsafeBuiltinPluginExtension<T extends "widget" | "block"> = {
+  type: T;
+  extensionId: string;
+  name: string;
+  component: FC;
+};
+
+export type UnsafeBuiltinWidgets<T = unknown> = Record<string, T>;
+
+export const loadUnsafeBuiltinPlugins = async (urls: string[]) => {
+  return (
+    await Promise.all(
+      urls.map(async url => {
+        try {
+          const plugin: UnsafeBuiltinPlugin = (await import(/* @vite-ignore */ url)).default;
+          return plugin;
+        } catch (e) {
+          throw new Error(`Specified unsafe built-in module could not find: ${url} ${e}`);
+        }
+      }),
+    )
+  ).filter(Boolean);
+};
